The project hasn't been set up, so the buildpath etc. is probably wrong.
The aggregatedsearch.php uses fsearchsites.xml to get the location of the ??_fs.xml. The folder fs has the current fs.xml files as seen from another box. The versions with the php files are set to work from "localhost".

fs.php is geared to reurn the vivo site results in the desired xml.
The ??_fs.xml sites point to the fs.php with the search data.

git-svn-id: http://svn.code.sf.net/p/vivo/tools/code@510 8d54b258-bb70-4df8-ae49-98ba5baa8db3